Add support for executeObject script and executeObject function.
Object scripts and Object functions are JS scripts or functions that
return a V8Object. The Object is automatically added to the global list
of objects and must be released by the caller.

If the execution of the script (or function) throws an exception, the
object is automatically released and the exception is thrown to the
caller.