Add support for IntFunction parameters.
Parameters are passed to functions as a V8Array. Each value in the array
is converted to a parameter in the JS Function. This adds support for
parameters to ExecuteIntFunction.

The V8Array that holds the parameters must be released after the
funciton is called, otherwise the resource will be leaked.

In addition to this, this change-set also provides support for Object
Functions, that is, functions declared on objects as well as the global
scope.