C++ Bindings for setPrototype
setPrototype sets the prototype on the V8Object. If applied to V8
itself, it will set the prototype of the Global Object. As in JS, after
a prototype is set, it can still be changed and those changes will be
reflected dynamically in the Object itself.

This is the C++ side of things.