Add debug support to J2V8
Debug Support allows the user to specify a debug port on which to attach
an external debugger. The API require the user to register a callback,
which will be invoked whenever a debugger is attached. It is the
responsibility of this callback to invoke V8.processDebugMessages() from
the proper V8 Thread.

Debug support can be registered with an option to halt the VM on the
first instruction and wait for a debugger to connect.