Store a global handle to the V8.class in our JNI code
Looking up the JNI class when a debug message fails on Android. This is
because the debug message happens on a separate thread and we don't have
a proper class loader at that point. Instead of looking it up then, we
now create a globabl reference to this class and store it. This is
created once, when the first V8 Isolate is created.