Add support for calling Java methods that return a V8Object.
When called from JS, Java method that return a V8Object return that
object to the calling JS function. The V8Object that is returned is
automatically released. This means that the Java code cannot use the
object it returns after the life the method. This is by design,
otherwise this would result in a memory leak.