A a tool to create socket tunnels
When debug is enabled, v8 only listens for connections from localhost.
However, in order to debug on android devices, we need to open these
connections to other hosts. The debug tunnel opens a public port and
forwards all the traffic to the specified port. This is a bi-directional
tunnel, so information can flow both ways.

This allows users to setup a public port and forward all traffic to the
private debug port on V8.