Make V8Arrays extends V8Objecst
In JS and in V8, Arrays are Objects. To make this usable in J2V8, the
V8Array class now extends the V8Object class. Furthermore, Objects and
Arrays are no longer stored separately in the native side registry.
Instead they are kept in the same registry.

Arrays elements can now be accessed using the get<Type>(key) method.

This fixes #9.