Add support for NamedScripts
Named Scripts are needed for Debug support so the user is shown a
meaningful name of their script when they debug. This has been
implemented in V8 using ScriptOrigins. The script origin is created in a
C style Macros to avoid having the same creation code in each execute
script method.

A backwards compatible API has been added that lets users run scripts
without a name (or starting line number). In this case, NULL is passed
for the ScriptOrigin and the V8 defaults are used.

This fixes #5