Add support for methods with unknown (late binding) return types
Return type of Object is common to allow for late binding. At execution
time, this might be a boxed primitive, or some other type. This is now
supported by looking at the runtime class and using the right accessor
to get the real value. We currently support Integer, Double, Boolean,
String, V8Array and V8List. If the return type is not one of these, then
a runtime exception is thrown.