Add support for conversion from Long and Float
We currently only support ints and doubles. However, in some cases Longs
and Floats are stored in Objects. With this change-set, V8ObjectUtils
will now handle these Objects, however, they will be converts to ints
and doubles. For the case of longs, this is a loss of percision. To
further support this, J2V8 core will need to be updated with long
support.