Support for missing parameters
In JS, functions can be called with missing paramters. When the function
is a registered Java function we use default values. For int the default
is 0, for double the default is 0, for boolean is the default is false,
and all other values the default is null.

This also supports methods with var arg parameters. Once the number of
non var-arg parameters has been reached, all the other parameters are
part of the var arg list.