Remove all the callback methods except void and object
Because we invoke all callbacks using reflection, we don't need a
separate method for each return type. With reflection you either call an
Object method or a void method. If it is an Object method, we determine
the type at execution time and apply that result to the V8 Runtime.