Add support for getting the type of an entire range in a V8Object
If a continous range of values in a V8Object have the same type,
getType(start, length) can be used to get this type. If the length of
the range is 0, or if it's out of bounds, or if all the types are not
the same then a V8ResultsUndefinedException is thrown.