Add a static method that can determine if the native libs are loaded
In order to use J2V8 the native libraries must be loaded. There are lots
of reasons why they may not be loadable. A new static method has been
added to the V8 class that determines if the libs are loaded.

This required some tricks in the tests. The tests had to be run in a
separate class loaded since the static block is only loaded once.