Add a utility method for geting Object values from Arrays
We used keys (string values) to get values from Arrays but this is not
as fast as using the index. This change-set introduces a utility method
to get the (object) value using an index. IF the Value is a V8Object it
is converted to a Map. If it's a V8Array, it's converted to List.