Add a utility method for getting an Object as a type J2V8 can use
GetV8Result resturns an object as a type V8 can use. Namely, if the
input is a Map, then a deep copy to a V8Object is done, and if the type
is a list then a deep copy to a V8Array is done. Otherwise the input is
just returned.