Support re-entrant script execution in J2V8
J2V8 did not support re-entrant script / function execution. That is,
from a Java callback, another call to JS would fail sporadically. This
is because the JNI context was setup on each execution, so the
re-entrant setup would destory the current one.

This was fixed by no longer setting up the context on each execution.
Instead, the pointer to the V8 Object is stored when the Runtime is
created and the JNI context is retreived from the JVM object when
needed.

Finally, checks were added to test if a released object is being
accessed.