Add support for ScriptExecution exceptions
Script execution exceptions are exceptions thrown by JS at execution
time. An exception can be thrown when either a JS Script is being
executed or a JS Function is being called. Futhermore, an exception can
be thrown from a Java callback, which is wrapped in a JS Script
Execution Exception and re-thrown.

ScriptExecutionExceptions contain information about the exception
including the message, line (and columuns) of code that caused the
exception and an optional stack trace.

This information has been encoded in a nice easy-to-read format and
returned as part of toString().