Provide an option not to report memory leaks
By default, J2V8 will report memory leaks when the runtime is released.
This is done by throwing an IllegalStateException. This is a good
default in debug mode, but arguably not the best behaviour once a
product has been released. This option allows developers to disable the
memory leak reporting.