Create a helper method for creating V8Strings.
To create a V8 String from a java string you first need to create a
cstring, then the v8 string, and then release the cstring. I helper
method has been created to perform these operations. The existing code
was refactored to use this method.

Futhermore v8:: was removed since we are already using the v8 namespace
and this is unnecessary.