When getObject returns a JS Array, it should be an instanceof V8Array
When getObject, or executeObjectScript / Function is called and a JS
Array is returned, the resulting value should be a V8Array, not just a
V8Object. However, since the Object is actually created before the JNI
Binding is called, this wasn't possible. The changes the design so the
actual object is created by the binding. This means the true type can be
returned. This will open the door for other types of V8Object creations
such as V8Functions and V8Dates.