Add an array getType() method
getType() will return the type of the array. That is, it will return the
type of all the elements of the array or UNDEFINED if they are not
compatible. If all the types are compatible (of the same type, or
subclasses of a common supertype) then that type is returned.