Allow JS to catch Java exceptions
When Exceptions are thrown from Java (during a callback) a JS
try / catch should catch these. However, the exception flag
on JNI was set to true, so catching these exceptions would not allow
any more calls to be made. 

With this change, we stash the exception and reset the flag. Only
the last exception will be stashed, so nesting Java exceptions
is not supported.

When the execute script call finally returns, if there any exceptions
left they will be propagated back to Java.