Create V8Object / V8Arrays on the native side for return values
When returning from a script / function, all return values for 
V8Object / V8Arrays are now created on the native side. This will
help when returning other types of things, such as undefined. Before
this, we used to create the objects on the Java side and initialize
them on the native side. However, this required unrolling if
anything failed on the native side.