Add the concept of undefined
When V8Objects or V8Arrays are accessed, but they are not available
an undefined type will be returned. This can be checked by 
using the isUndefined() method on the returned type. One undefined
result will be .equals() to all others.

Undefined types do not need to be released.

Calling any other method on an undefined type will result in an 
UnsupportedOperationException.