Remove the mapping between runtime handles and V8Runtimes
Instead of using the mapping between unique handles and V8Runtimes,
the pointer to the runtime is stored with the V8 Runtime Java Object.
This allowed us to remove the read-write locks since there is now
no shared datastructure to manage all the runtimes.

The Pointer to the runtime is casted to a 64bit long to be stored in
Java and back to a pointer when needed. 64Bit longs are used so it
works on 64 bit architectures.