Allow the V8 runtime to be compared against V8Objects
If the result of a V8 interaction returns a V8Object that represents
the global context, v8.equals(result) should return true. This did
not happen because V8 & V8Object were not the same class.

This fixes this. This also required a small change on the JNI side
to use the proper V8 Global Context within equals.

Finally, strictEquals and sameValue have been exposed on V8Values.