Add the receiver to the list of callback parameters
When calling a Java methods from JS, you may need to know the V8Object
that the function was called on. With this change, the V8Object (the
receiver) is now a parameter on JavaCallback#invoke and 
JavaVoidCallback#invoke. This is a breaking change from 2.2.

This parameter is automatically released when the flow of control
leaves the method scope. It should never be stored for future use.