Add a V8Executor to execute JS scripts on their own thread
A V8Executor can execute scripts on a separate thread in an isolated
runtime. The executor includes a terminate method to terminate
a long running script. A setup method can also be used to configure
the runtime.

The script to run can either be specified as a string, or returned
from a script loader (V8Script interface).