Add a Map utility to hold V8Values.
Often you need to keep V8Objects (V8Values) in Maps, but managing the
memory is a pain. This is because if a V8Object is stored as a key, it
can't be released. However, in some cases J2V8 will release those 
objects for you. To help with this, we should created a V8Map. A V8Map 
maps V8Values -> Objects. The values are twins, and the Map manages 
them. If they are removed, the map is responsible for releasing then.

This fixes #63.