Add the ability to register executors with the runtime that started them
Executors are often kicked off from an existing runtime, and when
that runtime shutdowns the executors that it started should also
terminate.

This change-set allows for executors to be registered with a runtime.
They can acquired and terminated. When the runtime is released,
all the executors are terminated and the list is cleared.