Break cycles when converting from V8Objects to Maps / Lists
When converting from V8Objects to Maps / Lists, cycles are now
broken. For example, if a map / list contains itself (either directly
or transitively), it won't lead to a stack overflow anymore.
A cache is kept as the conversion is done, and if a duplicate is 
reached, then the cached version will be used.