Use native memory address to reference V8 Objects
To provide a mapping between Java objects and native V8 Objects, we
now use the memory address of the native object. Historically we used
a UID that was generated from monotonically increasing counter.
However, this counter would eventually hit a limit even if all the 
native objects were released. Also, we had to maintain a map from
the UID to the native object.

Now the memory address of native object is used. The pointer is casted
to a jlong and stored in Java. This avoids the counter limitation and
also provides constant time look-up since each jlong can be casted
to a pointer and dereferenced.