Allow negative object handles for parameter lists
The parameter list passed to JSFunctions should allow both positive
and negative object handles. This is because a signed long is used
as a pointer location, which could be negative.

If no parameter array is used, then 0 will be passed as the
object handle.

This completes the work in 2330e326dcbef844569e2f5766167d90003844c9