Add support for a V8Error type.
V8Error types are java exceptions that are raised to JavaScript.
V8Errors contain a V8Object which is thrown once the control
is returned to JavaScript. The V8Error holds its own handle
to the V8Object. If control is returned to JavaScript, the handle
will be freed. If the exception is caught in Java before
control returns to JavaScript, the V8Error must be freed.