Add Java API for the V8 JavaScript debug API
V8 has a JavaScript debug API. This is the initial commit
towards a Java API for debug. The Java API will simply map
to the V8 JavaScript functions.

This change-set also adds support for setting flags on
the V8 Platform. This is required to use debug support
as the debug object must be exposed using:

--expose-debug-as=Debug

Change-Id: Ia1ac97438f14b0685756d8a2cb3e286e01de9301