Rework how V8 Functions are registered.
This change-set does two things, it allows unamed JS functions to
contain a body written in Java and it uses the memory address to
track method descriptors.

Instead of using global counters or UIDs to track the method descriptor,
we simply use the memory location. This means that the memory descriptor
ptr can be casted back to a memory descriptor and dereferenced.

Change-Id: I724e190849b79aa5c36049107edbb5b042c4fdf7