Use the native v8#get method to get the value of a property
A native implementation of Get was added for accessing any
object from V8. This is used in executeScript and executeFunction
where the return type is not known. This changes the implementation
of V8Object#get to use this instead of trying to determine the type
first. This method is heavily covered in other tests.

Change-Id: I44ba2e3dfa826324a07468f724fcd9bd0418d968