Use a single V8TypedArray to represent all typed arrays
Instead of using separate types for different typed arrays, we now
just use a single V8TypedArray. This is because not all array
types can be mapped to Java types. The type of the typed array
can be specified in the constructor using one of the V8Value
constants. Currently on V8Value.INTEGER and V8Value.BYTE are supported.
The V8TypedArray class supports a getByteBuffer to get the backing
store. This can be converted to specific views using the ByteBuffer
API.

With the new V8TypeArray support, V8Int32Array and V8Int8Array
have been removed.

Change-Id: If5d00de9e532d55d5008a46f3707542a41126b31