Add an Object persist API to the MemoryManager
When released, the MemoryManager will release all V8Objects created
while it was active. In order to keep some objects after the Memory
Manager has been released, a persist API was added. Objects that are
"persisted" will be removed from the MemoryManager's control.

In addition, the MemoryManager is not marked as 'released' unless all
objects are cleared without problems. If an exception is thrown
during release(), then the MemoryManager is considered still active.

Change-Id: I476b41c68a2155c34ebb378909a2952f61bd5d73