Protected against ArrayBuffers with no data
On some devices, we appear to be creating a NewDirectByteBuffer with
a null value. It's not clear how this happens, or why it only
happens on some devices. So if we know the result IS an ArrayBuffer
but the Data is null, just return NULL. This may move the problem
elsewhere, but since we can't reproduce this, we don't know.

Change-Id: I010db7b8e08e79722e4cc6818a2f96bb05a47573