Throw an IAE if null is used as a key in V8Object
If null is used as a key when accessing properties of a V8Object,
an IllegalArgumentException will be thrown. Before this change,
the entire VM would crash.

This is handled by checking the value of the key before any native
code is invoked, and fast failing if it's null.

This fixed #249

Change-Id: Ia6ff54f9586aed14179c0c6de1f7ef49d1730c67