Added an API to get the SCM Revision ID
To help determine which version of J2V8 is being executed
at Runtime, a new method was added on the V8 class.
This method will return a String representation of the
SCM Revision ID.

We currently use Git, so this method will return the
commit ID for the latest commit in a build.

It is expected that the default string will be replaced
when J2V8 is built for deployment.

Change-Id: I3a2e5f5f8f8cf88283d7ef1b1571257dab6a4fa3