Get the ScriptName from the Frame
V8 has changed the way the ScriptName must be accessed from the Frame.
Instead of assuming that the FrameMirror always as a Script Object
(which now does not appear to always be the case), we also
need to check if the FrameMirror has a FunctionMirror with a
script name attached.

In case both there is both a ScriptName attached to the FameMirror
and a ScriptName attached to the FunctionMirror, the one attached
to the FrameMirror is chosen.

Fixes #262

Change-Id: Ic6fca5e963bb205268045d33df487d775ea7954f