Add type adapter support to J2V8 conversion
V8ObjectUtils provides a number of conversion routines to convert from
V8 Types (V8Object and V8Array) to Java objects. This change-set
allows the user to influence the conversion by providing a type adapter
for each conversion. With the type adapter, you can change
how the types are converted, or use the default type adaption.

This change-set also adds a number of test cases for this and shows
how it can be used in practice.

Change-Id: I061128986ec4dc9f414424d46b2158e871ab9d39