Add support for getting the 'type' of a V8Value
Since this is implemented on V8Value, it only works for those types
that inherit V8Value. This method will return an integer, which maps
to one of the constants defined in V8Value.

Since V8Array already had a method called getType, we called this
method getV8Type.

Fixes #266

Change-Id: I5de9ceb3b223e5c35091ae9b68d687ec63625189