Add an API to convert V8 results to pure Java objects
We had separate APIs for converting V8 results to Maps and Lists,
but we didn't have a generic method for converting any result from
V8 to a Java object using a deep copy.

This includes a number of refactorings, because this method makes it
easier to perform the other deep copies in V8ObjectUtils.

Fixes #267

Change-Id: Id17424d94144104d40102c6f34a9f6e67a2f6a89