Add support for Weak references
In some cases we cannot track references entirely in Java and we
need Javascript to track the lifecycle of the Object. With setWeak
a V8Value can be marked as not needed in Java but will remain as
long as no more references exist to the Object. This allows
developers to use V8Values as long as they are still referenced
in Javascript.

V8 does not guarantee that these objects will every be released,
so Weak references should not used unless absolutely necessary.

Change-Id: I6507e2619f8212b65d4624e1dedff0b84724e355