Adds a method to check if an object has been set weak
Uses V8 to determine if setWeak has been called on this Object. This
change-set also changes the way `setWeak` works. `setWeak` now returns
this, so object chaining can be used.

Finally, V8#referenceCounts have been updated to include a count
of Weak objects. That is, if an object is marked weak, it should
not count towards the Object count.

Change-Id: I28d9023add2ff7e180bd30356665b9055df9cd38