Do not "re"-acquire the native lock
Locks in V8 are counting semaphores, but J2V8 uses a simple binary
semaphore. This means if that the same lock is acquired multiple
times, and released once, V8 will still consider the lock
"acquired" but J2V8 won't. This leads to deadlocks.

To fix this, we don't ever acquire a native lock more than once. If
it's already acquired, we simply return.

This fixes #277

Change-Id: I420ecb3e45a3931e592b10d5bb5782ecf1e64d1b