Get existing build scripts working with updated build
When the build was updated to use CMake, the existing build
scripts stopped working. Ideally we will remove these old
scripts, but they are still being used in a number of places.

The pom.xml now has an environment variable that is used to
specify the version and name. We now set these in the buildAll.sh
script.

Change-Id: Ia1e6974ba2092568f0918fa667907107b1d2f5ba