Allow calls to V8Object#toString() on released objects
If a V8Object has been released, calls to toString() will no
longer fail. If the object was an Object, then [Object released] is
returned. If the Object is a Function, then [Function released] is
returned. If the Object is an Array, then [Array released] is
returned.

Also, if the runtime has been released, an someone still has a handle
to the object, the toString() method also works with the same semantics.

Change-Id: I7ead272213c65e3fde29445c9c3e730bf350d7a0