Do not create any scopes dynamically
It is important that all Enter() be coupled with an Exit(), otherwise
the current scope in V8 may not be what you expect. We accomplish
this by managing our Enter / Exit through Isolate::Scope. When
these are stack based, we can be sure that C++ scopes will handle
this for us. However, when they are heap based, we cannot be sure.

All our Scopes are stack based except for an isolate scope we create
when the isolate is first created. This is deleted when the isolate
is released. However, this means that if an isolate is created on
the same thread as another isolate, V8 may not know which is the
current one. To fix this, we now make sure all Scopes as stack based.
All JNI calls will create a scope, and that scope will be removed
when the call from JNI returns. That way, all JNI calls explicitly
state their scope.

This fixed #321

Change-Id: I99be8c08812d56d118a0de1f7f1352f8f2b58d52