Throw an exception if a lock is released while in a context
The V8Locker cannot be released in a callback, and must only
be released when no J2V8 API is currently on the stack. This is
because we currently have a context, and if we release, we cannot
properly lock again.

To protect against this, throw an exception if release or acquire
is called while the isolate is in a context.

This fixed #318

Change-Id: Iacde6ca55ad368b7e4f5b585eea389bffd0ca082