Remove the text relocation from the build
The x86 binaries were failing at runtime on Android because of
text relocation that existed in the shared object. The text
relocation was a result of the openssl library. This libary
is not needed for J2V8 so it was removed.

This change-set also updates the NDK to version 14 and updates
the version of CMake that is used during the build. With
this change-set applied, x86 and ARM builds should now work.

The version was set to 5.1.0-SNAPSHOT, as we had lots of failed
builds during 5.0.

Change-Id: I1bef1e0226ddf2e209b48d3f6467bb287acc421f