Create a gitlab ci build for J2V8
Add a gitlab-ci yaml file so J2V8 can be built on GitLab.
The gitlabci uses a base docker-dind image, and adds the
required dependencies. It also sets the commit hash
in the V8 method that returns the SCM ID. This is useful
when debugging.

In addition to the new build, this fixes a typo in the name
of the emulator. Since we updated to a newer NDK build, the emulator
was suffixed with '64'.

Change-Id: I60b420d16ca26f06a4046b9793e31c47de5220cd