Fail the build if tests are failing
If the build output contains the words BUILD FAILED then
we fail the build on gitlab. In the past it was passing.
Because the build runs in a contained bash env, running on
supervisord, in a container, it was not possiible to check
the status code of gradle. Instead we check the logs.