Skip OS Specific tests when running on the emulator
Many of the platform detector tests fail when running
on the emulator. This is because the test attempts to
set the OS property to something else which is not allowed.
Instead, we skip these tests as it doesn't make sense
to try and assume that the android emulator is
Windows, Linux or Mac.